introduce-raw-tokenizer-wrapper-move: Hand tokenizer/processor ownership over to RawTokenizerWrapper
Pure cut/paste move per MECH_COMMIT_SPLIT. Cuts:

  - ``TokenizerManager.init_tokenizer_and_processor`` (@staticmethod
    in prep) → ``RawTokenizerWrapper.init_tokenizer_and_processor``
    (instance method, drop @staticmethod, restore plain ``self``)
  - module-level helpers ``_get_processor_wrapper`` and
    ``_determine_tensor_transport_mode`` from tokenizer_manager.py →
    raw_tokenizer_wrapper.py module top-level

Updates imports in both files (move what the body needs, drop what TM
no longer uses). No external caller rewrites — the @property facade on
TokenizerManager (added in prep) absorbs every ``tm.tokenizer`` /
``tm.processor`` / ``tm.mm_processor`` access. InputFormat enum + the
4 tokenize-pipeline helpers move in the follow-up
``rtw-move-tokenize-helpers`` commit.

Refactor chain ID: introduce-raw-tokenizer-wrapper-move
